Update NeMo_TTS_Primer.ipynb (#6436)
* Update NeMo_TTS_Primer.ipynb

Changed a mistake in line 782. Instead of frequency band (ie. pitch) we should write frequency bin. Note that frequency bins in FFT are not related to pitch.

Signed-off-by: Mostafa Ghorbandoost <mos.ghorbandoost@gmail.com>

* Update NeMo_TTS_Primer.ipynb

Corrected the description of spectrogram and mel spectrogram calculations in lines 782 & 783 and added a fourth point to the description and added a reference for more mathematical details at the end of this point.

Signed-off-by: Mostafa Ghorbandoost <mos.ghorbandoost@gmail.com>

---------

Signed-off-by: Mostafa Ghorbandoost <mos.ghorbandoost@gmail.com>