Allow inference without reference audio
In some cases we need to run inference on manifests that do not include context audio
and/or ground truth audio:

* Text context manifests may not have a context audio included (and it wouldn't be very
relevant anyway)
* When generating from arbitrary text the ground truth audio cannot be assumed to be
available. Note that even without the context and ground truth there are useful metrics
that can be calculated like WER, UTMOS, and inference speed.

This commit modified the inference and dependent scripts to allow the absence of context
and/or ground truth audio.

Signed-off-by: Fejgin, Roy <rfejgin@nvidia.com>